fix (notification-service): Correct Kafka message format

+++ backend/notification-service/src/main/java/com/github/spjavaind300/notificationservice/service/AdvService.java
@@ -4,6 +4,6 @@ import com.github.spjavaind300.notificationservice.model.dto.AdvDto;
 
 public interface AdvService {
 
-    AdvDto getAdv(long advId);
+    AdvDto getAdv(int advId);
 
 }
